+++ src/sct/cli/common.py
@@ -31,7 +31,6 @@ input_point_target_option = click.option(
     "--point-target-source",
     "-pt",
     required=True,
-    default=None,
     type=click.Path(path_type=Path, exists=True, dir_okay=True),
     help="Path to the external point target source",
 )