+++ src/sct/cli/sct_cli.py
@@ -14,7 +14,7 @@ from sct.__init__ import __version__ as VERSION
 from sct.cli.int_cli import interf_coherence_analysis
 from sct.cli.pta_cli import target_analysis
 from sct.cli.ra_cli import radiometric_analysis
-from sct.cli.sa_cli import spectral_pt_analysis
+from sct.cli.sa_cli import spectral_analysis
 from sct.cli.tar_cli import pt_ambiguity_ratio_analysis
 from sct.configuration.logger import sct_logger
 from sct.configuration.sct_configuration import SCTConfiguration
@@ -48,5 +48,5 @@ def sct_analysis(ctx: click.Context, config: Path | None) -> None:
 sct_analysis.add_command(target_analysis)
 sct_analysis.add_command(radiometric_analysis)
 sct_analysis.add_command(interf_coherence_analysis)
-sct_analysis.add_command(spectral_pt_analysis)
+sct_analysis.add_command(spectral_analysis)
 sct_analysis.add_command(pt_ambiguity_ratio_analysis)